requirements.txt
amqp==5.0.6
anyjson==0.2.5
asgiref==3.4.1
bcrypt==3.2.0
billiard==3.6.4.0
bleach==3.3.0
celery==5.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==3.3.1
Django==3.2.10
django-bleach==1.0.0
django-braces==1.15.0
django-choices==1.7.2
django-ckeditor==6.0.0
django-cors-headers==3.10.1
django-countries==7.2
django-crispy-forms==1.11.0
django-currentuser==0.5.3
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
django-js-asset==1.2.2
django-loginas==0.3.9
django-nested-inline-bug-fixed==0.4.4
django-phonenumber-field==5.0.0
django-postgrespool2==2.0.1
django-recaptcha==2.0.6
django-redis==4.12.1
django-tabbed-admin==1.0.4
django-wkhtmltopdf==3.3.0
docopt==0.6.2
Fabric3==1.14.post1
Faker==10.0.0
idna==2.10
kombu==5.2.1
mixer==7.1.2
packaging==20.9
paramiko==2.7.2
phonenumbers==8.12.17
Pillow==8.4.0
pipreqs==0.4.11
psycopg2-binary==2.9.2
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
text-unidecode==1.3
urllib3==1.26.3
webencodings==0.5.1
yarg==0.1.9
